refactor: extract PublicApiSurfaceExtractor shared helper
